Consolidate draft persistence into one file
Track B structural findings. Behavior-preserving.

- New WorkDraftPersistence.swift owns both draft stores, the storage mechanism
  they share, and the composer-draft modifier. WorkQuestionDraftStore had been
  living in a View file while its identically-shaped sibling lived in
  WorkModels.swift, which is how their duplication stayed invisible; both files
  drop back under the 1k-line threshold.
- Extract WorkDefaultsJSONMap (load / persist / evictingOldest) as three free
  functions. The two stores agree on the mechanism but not the policy — key,
  cap, and what counts as a no-op write all genuinely differ — so a shared
  store type would only have to model those differences back out.
- Move questionDrafts' updatedAt out of Snapshot into a Stored wrapper, so the
  "did the answer change?" check no longer has to neutralize the timestamp on
  a mutable copy first. Key bumped to v2; open-gate drafts are ephemeral.
- Replace the hand-copied restore/debounce/flush blocks in the Hub and New Chat
  composers with .workPersistedDraft(_:key:), and share the 400ms debounce with
  the two surfaces that schedule their own.
- Gate the question card's keyboard Done toolbar on freeformFocused. Keyboard
  toolbars scope to the enclosing view and this card sits just above the main
  composer's UITextView, which that button cannot dismiss — a leaked Done that
  silently does nothing is worse than no Done at all.
