+++ apps/ios/ADE/Views/Hub/HubComposerDrawer.swift
@@ -274,22 +274,7 @@ struct HubInlineComposer: View {
         }
     )
     .onAppear { onAppearSetup() }
-    // Restore whatever the user last typed here but never sent. Guarded on
-    // empty so a re-appear (or an init-seeded value) can't clobber live text.
-    .task {
-      if draft.isEmpty {
-        draft = WorkComposerDraftStore.load(WorkComposerDraftStore.hubNewChatKey)
-      }
-    }
-    // Debounced autosave: each keystroke restarts this task, and the cancelled
-    // sleep throws before the write, so only a typing pause hits UserDefaults.
-    .task(id: draft) {
-      try? await Task.sleep(for: .milliseconds(400))
-      guard !Task.isCancelled else { return }
-      WorkComposerDraftStore.save(draft, for: WorkComposerDraftStore.hubNewChatKey)
-    }
-    // The debounce dies with the view, so flush the final text on teardown.
-    .onDisappear { WorkComposerDraftStore.save(draft, for: WorkComposerDraftStore.hubNewChatKey) }
+    .workPersistedDraft($draft, key: WorkComposerDraftStore.hubNewChatKey)
     .onChange(of: composerFocused) { _, focused in
       if focused { withAnimation(hubComposerSpring) { expanded = true } }
     }
