+++ apps/ios/ADE/Views/Work/WorkModels.swift
@@ -64,93 +64,6 @@ enum WorkComposerPreferences {
   }
 }
 
-/// Unsent composer text, persisted per surface so leaving a chat (or the app)
-/// never discards what the user typed — desktop keeps its draft, and mobile
-/// silently dropping it was the single most-reported chat regression.
-/// One JSON dictionary under a versioned key: small enough to rewrite whole on
-/// each save, bounded by `maxEntries` (LRU by `updatedAt`) so a long-lived
-/// install can't grow it without limit.
-enum WorkComposerDraftStore {
-  struct Entry: Codable, Equatable {
-    var text: String
-    var updatedAt: Double
-  }
-
-  /// Versioned so a future shape change can migrate rather than mis-decode.
-  private static let storageKey = "ade.work.composerDrafts.v1"
-  /// Enough to cover every chat a user realistically juggles; older drafts are
-  /// evicted oldest-first rather than kept forever.
-  private static let maxEntries = 60
-  /// A composer draft is a prompt, not a document — clamp pathological pastes so
-  /// one entry can't dominate the shared defaults store.
-  private static let maxLength = 20_000
-  private static var defaults: UserDefaults { ADESharedContainer.defaults }
-
-  /// Per-chat key. Blank session ids yield a blank key so callers that render
-  /// before the session resolves can't write everyone's draft into one bucket.
-  static func chatKey(sessionId: String) -> String {
-    let trimmed = sessionId.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "" }
-    return "chat:\(trimmed)"
-  }
-
-  /// The two "new chat" composers are singletons, so they get fixed keys.
-  static let hubNewChatKey = "hub-new-chat"
-  static let workNewChatKey = "work-new-chat"
-
-  /// The stored draft, or "" when the key is blank, absent, or undecodable —
-  /// restoring must never be able to fail loudly in a view body.
-  static func load(_ key: String) -> String {
-    guard !key.isEmpty else { return "" }
-    return loadAll()[key]?.text ?? ""
-  }
-
-  /// Persists (or clears) the draft for one surface. An emptied composer removes
-  /// its entry outright: a user who deletes their text must not have it
-  /// resurrected the next time the screen mounts.
-  static func save(_ text: String, for key: String) {
-    guard !key.isEmpty else { return }
-    var map = loadAll()
-    guard !text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty else {
-      guard map.removeValue(forKey: key) != nil else { return }
-      persist(map)
-      return
-    }
-    let clipped = String(text.prefix(maxLength))
-    // Autosave runs on a keystroke debounce; skip the UserDefaults write when
-    // the content is unchanged so idle typing pauses cost nothing.
-    if map[key]?.text == clipped { return }
-    map[key] = Entry(text: clipped, updatedAt: Date().timeIntervalSince1970)
-    if map.count > maxEntries {
-      let survivors = map
-        .sorted { $0.value.updatedAt > $1.value.updatedAt }
-        .prefix(maxEntries)
-      map = Dictionary(uniqueKeysWithValues: survivors.map { ($0.key, $0.value) })
-    }
-    persist(map)
-  }
-
-  /// Drops a draft that has been consumed (sent) so it can't reappear.
-  static func clear(_ key: String) {
-    guard !key.isEmpty else { return }
-    var map = loadAll()
-    guard map.removeValue(forKey: key) != nil else { return }
-    persist(map)
-  }
-
-  private static func loadAll() -> [String: Entry] {
-    guard let data = defaults.data(forKey: storageKey),
-          let decoded = try? JSONDecoder().decode([String: Entry].self, from: data)
-    else { return [:] }
-    return decoded
-  }
-
-  private static func persist(_ map: [String: Entry]) {
-    guard let data = try? JSONEncoder().encode(map) else { return }
-    defaults.set(data, forKey: storageKey)
-  }
-}
-
 enum WorkToolCardStatus: String, Equatable {
   case running
   case completed
